Start the test in the right state
We'll set up a ticket or two to explore this issue. I don't think that the component test is the right place for verifying that the app starts in the right state and we'll probably move that down the line to maybe Application tests (depending on some details about how the tests isolate window variables) but probably Selenium tests.