fix: footer at the bottom w/o limiting body
Fix an error introduced in #624.

Using percentage values to make the application div fill the viewport
creates dependencies from viewport to <html> to <body> to app div.
This constraint on <body> (always the same height as the viewport)
caused the issue solved by #624, but that fix broke the "footer at
the bottom" behavior.

Instead of all that, use `vh` to more directly express what we want:
The application element should be at least as tall as the viewport.