Add colors and css classes to allow provider descriptions more color options (#1181)
- Ticket: [ENG-2751](https://openscience.atlassian.net/browse/ENG-2751)
- Feature flag: n/a

## Purpose

Give product the ability to do limited changes of the color of the Provider Description text and links. One would use it by wrapping the description in a span with the appropriate classes. For example, `<span class="ColorBlack LinkColorPurple">This is the provider description <a href="https://osf.io/">with a link</a>.</span>`. I added the following CSS classes for use by Product:

For text: 

- ColorBlack
- ColorPureBlack
- ColorBlue
- ColorPureBlue
- ColorGrey
- ColorRed
- ColorGreen
- ColorYellow
- ColorTurquoise
- ColorPurple

For links:

- LinkColorBlack
- LinkColorBlue
- LinkColorPureBlue
- LinkColorGrey
- LinkColorRed
- LinkColorGreen
- LinkColorYellow
- LinkColorTurquoise
- LinkColorPurple

## Summary of Changes

1. Add some colors to the _variables.scss
2. Add global classes, restricted to the local scope, that will change the color of the text and links of the provider description.

## Side Effects

Shouldn't be. The tooling that our admin app provides will strip out inline css, so we had to provide classes that would be usable from a dynamic text field in order to make color changes to the text and links. This meant using global classes, but I kept those classes only active within the ProviderDescription local class namespace. So everything should be restricted.

## QA Notes

This will only affect the discover page of a branded provider that's using the Provider Description option from the admin app. Instructions for use are at the top of this PR in the Description section.