Remove the VOL from the currentUser when stopping view only (#1278)
-   Ticket: [ENG-1578](https://openscience.atlassian.net/browse/ENG-1578)
-   Feature flag: n/a

## Purpose

The existing functionality of 1578 is only half-fixed. It's removing the VOL from the URL, but leaving it on the currentUser service, so the VOL banner still shows on the dashboard, and transitioning to any other route will put the VOL back on the url. This fixes it so that the VOL is cleared from the currentUser as well.

## Summary of Changes

1. Clear the VOL from the currentUser service.

## Side Effects

Nope.

## QA Notes

This should just be the same test plan you used before for 1578.