Ensure logout action always takes the user to the home page (#1272)
-   Ticket: [ENG-3027](https://openscience.atlassian.net/browse/ENG-3027)
-   Feature flag: n/a

## Purpose

Ensure logout always takes user to the logged-out home page.

## Summary of Changes

1. Pass the home page url to the call to _authRedirect on invalidationSucceeded.

## Side Effects

Shouldn't be.

## QA Notes

Everywhere in the ember-osf-web application should now redirect to home on logout. There's no reason this should affect anything else in the app that I can think of. Preprints already has this behavior, so we'd just need the legacy app to also do this to be completely consistent across the board.