Replace bstabs on project-registrations page with ember-aria-tabs (#1271)
-   Ticket: [ENG-3159](https://openscience.atlassian.net/browse/ENG-3159)
-   Feature flag: n/a

## Purpose

Fix a11y problems on the project/registrations page's tabs by replacing them with better tabs.

## Summary of Changes

1. Add ember-aria-tabs to project
2. Replace BSTabs with ember-aria-tabs on the project/registrations page
3. Fix test selectors
4. Remove unused tab tracking code
5. Ensure proper tracking information is sent to GA
6. Fix styling

## Side Effects

The color on the inactive tab changed from blue to black because on BSTabs the tabs are `<a>` tags, but they aren't really functioning as links, so I didn't carry over the color scheme.

## QA Notes

This ticket just fixes the tabs. The other issue from this ticket was covered by https://github.com/CenterForOpenScience/ember-osf-web/pull/1268

Also, there is a lot of functionality in the vicinity of these tabs. Nothing should be broken as it's fairly well tested in our code (and I poked around at it as well), but just so you know.