[ENG-3206] No link if no wiki (#1283)
-   Ticket: [ENG-3206](https://openscience.atlassian.net/browse/ENG-3206)
-   Feature flag: n/a

## Purpose

Hide the wiki link if there's no wiki on a registration so that we don't have a bad link on the page

## Summary of Changes

1. Add wikiEnabled to model and mirage
2. Use wikiEnabled to determine if we should show the wiki link
3. Test

## Side Effects

No, this is fairly isolated

## QA Notes

To test, make a project, disable its wiki in the settings, and make a registration of that project. Then visit the registration's page to verify that there's no wiki link. Also verify that registrations with wikis have wiki links.