[ENG-3225] Get provider types instead of all registration types (#1302)
-   Ticket: [ENG-3225](https://openscience.atlassian.net/browse/ENG-3225)
-   Feature flag: n/a

## Purpose

Instead of grabbing the top-level schemas, grab only the provider schemas for the search facet.

## Summary of Changes

Aside from just using the provider endpoint, I had to get the OSF provider if we were on a page that doesn't have one. As the component is currently used, there will never be a provider, because it's only visible on the roll-up discover page, and only when the OSF Registries registry is a selected facet. However, there is likely going to be a request in the near future to allow faceting for providers, and the easiest way to do that will be to show the component on provider-specific discover pages. So I made it easy to work with that if we go that direction, otherwise there'll be a little logic in here that's more complicated than it needs to be.

## Side Effects

The analytics tracking is going to include the OSF registry provider where it did not before, because there's always going to be a provider now.

## QA Notes

This component should only show the schema types that are listed in the admin as being valid for OSF Registries (well, plus the ERPC schema type).