[No ticket] Cache providerIsDefault to avoid problem where this.provider is undefined (#1339)
-   Ticket: n/a
-   Feature flag: n/a

## Purpose

When doing a search from the registries landing page, we would get an error on the discover page that we couldn't load the registration types. This happened because provider was becoming undefined after the await L45 in the diff for this PR was executed.

## Summary of Changes

1. Wait for attrs to load before running task (h/t @aaxelb)

## Side Effects

This should be pretty isolated

## QA Notes

Original error was when performing a search from the registries landing page, the error above would show on the discover page. This should no longer happen.