[ENG-3300] Properly encode provider name in GA event (#1343)
-   Ticket: [ENG-3300](https://openscience.atlassian.net/browse/ENG-3300)
-   Feature flag: n/a

## Purpose

We were sending the string '@provider.name' to Google Analytics instead of the provider's name in a few instances. This corrects that.

## Summary of Changes

1. Use @provider.name in registry discovery results
2. Use @provider.name in sort dropdown
3. Use @provider.name in sharing control

## Side Effects

No, this is isolated.

## QA Notes

Should just need to check the `collect` network request to ensure that the provider name is being included in the request for analytics information collected on a branded provider's discover page.