[ENG-3357] Make hover state lighter and more contrasty (#1362)
-   Ticket: [ENG-3357](https://openscience.atlassian.net/browse/ENG-3357)
-   Feature flag: n/a

## Purpose

The external services links on the registries landing page were too dark compared to their background in the hover state. This lightens up the hover state and increases the contrast.

## Summary of Changes

1. Add hover state css.
2. Fix the link for "open source code" to go to ember-osf-web instead of ember-osf-registries.

## Side Effects

This is isolated to those links and the hover state.

## QA Notes

Just verify that I made the hover state light enough to be properly visible.