[ENG-3585] Filespage/fix download provider (#1406)
-   Ticket: [ENG-3585](https://openscience.atlassian.net/browse/ENG-3585)
-   Feature flag: n/a

## Purpose

Root folders are incomplete and buggy in the OSF API. For OSF Storage, they provide a url that is incompatible with waterbutler. For other storage providers, it's not implemented. So, instead of using the root folder, this PR uses the storage provider's files relationship, which provides the same info in a way that's complete and functional across providers. This should also allow the download provider as zip functionality to work.

## Summary of Changes

1. Add mirage for provider files relationship
2. Add provider files relationship to model
3. Create provider-file and osfstorage-provider-file
4. Make osfstorage manager use the osfstorage-provider-file when appropriate
5. Fix download as zip test
6. Don't embed region, because the osf api *hates* it

## Side Effects

Shouldn't be side effects. This should function just as it did before, just better. 

## QA Notes

Nothing different from what should have been happening all along.