[ENG-3811] Make it easier to select rows (#1547)
## Purpose

When I added the file sizes, I made the flex box act columnar. This had the side-effect of extending the divs/buttons that had a z-index of 9, making the click target for a selection much smaller than it had been. This reduces the surface area of the z-indexed items to just be the clickable targets and prevents the button from extending the width of the column in the case of folders.

## Summary of Changes

1. Juggle around z-index
2. Restrict size of button to not be the width of the column