[ENG-3811] Fix storage limit code (#1550)
## Purpose

Make storage usage actually work in the places it needs to. The basic problem was that, while getters get magic from ember to wait for async calls as part of .get(), they don't get access to that if the gets are in another async function. So I moved the getters for the storage limit into the getters for the properties that need that info.

## Summary of Changes

1. Remove async helper function
2. Update `userCanMoveToHere`
3. Fix incorrect property on multi-select move button
4. Add 'withStorage' trait where needed

## Side Effects

You have to use the 'withStorage' trait if you want a mirage node that can use its files page.