[ENG-3915] Modify overview leftnav to show badges (#1602)
## Purpose

The purpose of these changes were to update the left hand navigation with properly rendered resource badge icons on the registries overview page. 

## Summary of Changes

Changes include the creation of a resource component invoked on the overview page with parameters of the current registration's resources. These include active and inactive versions of the badges whose display is determined by the resource's `finalized` property. Because badges displayed here are read-only, permissions based logic has not been added.