Don't try to filter on subjects - not implemented (#1645)
-   Ticket: [ENG-2293]

## Purpose

Don't try to filter on subjects. It's not fully implemented, so it causes a particular bug when the query parameter is in the url. This is a minimal-change fix to this to keep the search page from being stuck on zero search results, rather than finishing out the feature or deleting it entirely.

## Summary of Changes

1. Remove the subjects filter from the list of filters. (h/t @futa-ikeda)