Add missing services to fix linked projects list (#1646)
-   Ticket: [ENG-3511]

## Purpose

The linked projects list in the project analytics engine wasn't rendering. It looks like, sometime after we implemented the list, we added the toast and current-user services to the node-card (or something in the node-card chain). Those services weren't in the analytics engine, so the rendering of the node-card broke. This corrects that problem.

## Summary of Changes

1. Add missing services to engine