Make dialog box scroll on overflow (#1649)
## Purpose

The modal for the links on the project analytics page doesn't scroll. Instead, it just runs off the screen if there's too much content, and that content will never be seen. This makes it so that the dialog box behaves properly with the scrolling and the not overfilling the screen.

## Summary of Changes

1. Add max-height to the modal
2. Set overflow to scroll

## Side Effects

If there were any other `bs-modal-simple`s on the page, they would also be upgraded. But there aren't, so this is isolated.