[ENG-4157] Add mirage for osf api (#1701)
-   Ticket: [ENG-4157]

## Purpose

Provide mirage endpoints for the OSF parts of the metadata project. The way this works is that, any time a guid is created, the appropriate custom-metadata record will also be created to go with. This means that if you navigate to, for example, a registration via the my-registrations page, a metadata record will not be created, because it won't go via the guid route, and a guid won't be created for it.

We _could_ automatically create guids for all records, but I think that would be a bad idea overall. It would make two related records for anything that we have that's a guid-based object, and that would likely slow down mirage a bit. It's probably best to just know that you may have to reload the page to get your metadata record when you're testing metadata things.

## Summary of Changes

1. Add factories for the custom metadata records
2. Add a factory for Guids to automatically add metadata records to the guid
3. Make the guid creation use mirage so that it activates the factory
4. Add endpoints to the mriage config
5. Serialize the `custom_metadata` relationship in the guid
6. Add serializers for custom metadata records to fix `self` link and add `guid` relationship