[ENG-4056] Guid Metadata (#1782)
-   Ticket: [ENG-4056]
-   Feature flag: n/a

## Purpose

There is tremendous value in digital objects having metadata to describe them. Metadata is a set of data that describes and gives information about other data. The OSF already has good practices for metadata on projects, registrations and preprints.
Limited metadata is currently provided about files on the OSF. 

## Proposed solution

Users have requested the ability to provide additional context about their files, in the form of metadata. We should start with a constrained set of metadata that can be applied to all files, projects/components, and registrations. To accomplish this, we are adding a flexible metadata relationship to all GUIDs in the OSF, which will be applied to start with on files, projects/components, and registrations. We will not apply the new metadata to preprints or users for now. In the future, we may wish to support additional discipline-specific schemas. File metadata should be designed and implemented in line with the FAIR principles.

[ENG-4056]: https://openscience.atlassian.net/browse/ENG-4056?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ