[ENG-4024] Deprecation/legacy test helpers (#1821)
## Purpose

Unsilence the legacy test helpers using the [new test api for engines](https://github.com/ember-engines/ember-engines/pull/653).

## Summary of Changes

1. Use new test api
2. Fix misc problems

## Side Effects

The test server seems to not want to rebuild after making changes to tests. Hopefully this isn't a larger problem with the rebuild process, or if it is, that it goes away after further upgrades. Also one test broke in a way that I can't find a reason for, so I'm hoping that an update to ember-qunit and ember-qunit-sinon will fix. A ticket has been made to do that, but it'll have to wait until after we've gotten rid of ember-bootstrap.