[ENG-4417] Remove BsNav (#1844)
## Purpose

Remove all usages of BsNav from the ember-osf-web codebase.

## Summary of Changes

1. Remove bs-nav from settings page
2. Remove bs-nav from registries moderation
3. Remove bs-nav from collections moderation
4. Add some data-test selectors

## Side Effects

In the previous version of the settings page, the active element was not highlighting the same way as it does on osf.io legacy pages (it wasn't highlighting at all, in fact). This version does highlight like legacy, but not for Firefox. The test selector I'm using works everywhere but there. So it's better and not worse, but it's not perfect. I think, since we went five years with nobody noticing, it's probably fine.