[ENG-4687] Add v2 models, adapters, and serializers (#2045)
Purpose

Add models, adapters, and serializers for various v2 endpoints we'll need for addons. Note that some of this does not reflect the BE as it exists today, but is a bit aspirational. That being said, it does not include all the changes we ultimately want, just the ones that are necessary to make this function in a sane manner.

Summary of Changes

Add addon, external-account, node-addon, and user-addon models, adapters, and serializers