Fixed the file upload issue (#2552)
-   Ticket: ### [ENG-7588]
-   Feature flag: n/a

## Purpose
There was the slightest chance of a race condition that I fixed using the "preUpload" called and a spinner.

I first tried to use "enabled=false" and it literally destroys the entire widget. So then I moved to the spinner

## Summary of Changes
Add a variable and preUpload method to the component.
Added the variable to the spinner conditional statement to display the spinner once the file is uploaded.

[ENG-7588]: https://openscience.atlassian.net/browse/ENG-7588